+++ eiam-20211201/alibabacloud_eiam20211201/models/_list_applications_response_body.py
@@ -66,6 +66,7 @@ class ListApplicationsResponseBodyApplications(DaraModel):
         self,
         application_creation_type: str = None,
         application_id: str = None,
+        application_identity_type: str = None,
         application_name: str = None,
         application_source_type: str = None,
         application_template_id: str = None,
@@ -87,6 +88,7 @@ class ListApplicationsResponseBodyApplications(DaraModel):
         self.application_creation_type = application_creation_type
         # The ID of the application.
         self.application_id = application_id
+        self.application_identity_type = application_identity_type
         # The name of the application.
         self.application_name = application_name
         # The origin of the application. Valid values:
@@ -152,6 +154,9 @@ class ListApplicationsResponseBodyApplications(DaraModel):
         if self.application_id is not None:
             result['ApplicationId'] = self.application_id
 
+        if self.application_identity_type is not None:
+            result['ApplicationIdentityType'] = self.application_identity_type
+
         if self.application_name is not None:
             result['ApplicationName'] = self.application_name
 
@@ -212,6 +217,9 @@ class ListApplicationsResponseBodyApplications(DaraModel):
 
         if m.get('ApplicationId') is not None:
             self.application_id = m.get('ApplicationId')
+
+        if m.get('ApplicationIdentityType') is not None:
+            self.application_identity_type = m.get('ApplicationIdentityType')
 
         if m.get('ApplicationName') is not None:
             self.application_name = m.get('ApplicationName')