+++ eiam-20211201/alibabacloud_eiam20211201/models/_update_cloud_account_description_request.py
@@ -18,7 +18,6 @@ class UpdateCloudAccountDescriptionRequest(DaraModel):
         # 
         # This parameter is required.
         self.cloud_account_id = cloud_account_id
-        # This parameter is required.
         self.description = description
         # IDaaS EIAM实例的ID。
         #