+++ eiam-20211201/alibabacloud_eiam20211201/models/_update_cloud_account_role_description_request.py
@@ -22,8 +22,6 @@ class UpdateCloudAccountRoleDescriptionRequest(DaraModel):
         # This parameter is required.
         self.cloud_account_role_id = cloud_account_role_id
         # 描述
-        # 
-        # This parameter is required.
         self.description = description
         # IDaaS EIAM实例的ID。
         #